mini-dec: make most functions return a struct str
We need both the pointer and the length at many places. A few tests
have shown that gcc is very efficient at returning a struct made of
two longs on x86 as it uses two registers for this. This is exactly
the situation we have here, so instead of having to write complex
and inefficient code, use a struct on all string outputs. Tests on
ARM show that the stack is used, but still efficiently.