WIP/mini-dec: reimplement a simpler and more memory-efficient model
This one relies on the in-order deletion to simplify everything. The
code is smaller, less memory areas are touched each time, the overhead
per table is 16 bytes, plus 8 bytes per cell, meaning that up to 3kB
can be saved on a full 4kB table (128 empty headers). On the average
29.2 bytes per header (66 headers), 1.5kB are still available in the
table when the protocol considers it's full, so the likeliness to fill
it and to have to defragment it is quite low.

Some more tests have to be run, and the defragmentation has to be
implemented. Likely a test will have to be implemented to stop going
upfront when there's little room left before the DTE table to avoid
realigning the table when there's a lot of room left at the end.
For this we could check if according to the protocol we're supposed
to have enough data or not, so that we know we won't fill the table.