mini-dec: implement dynamic table defragmentation
New function dht_defrag() defragments an existing dht by copying it
entirely into a new one. For now the new one is then copied again over
the original one for easier processing but it should instead return a
distinct one once the code is ready to accept this.

Defragmentation happens very rarely, once every few thousands of header
fields, so it's not worth wasting CPU cycles trying to be smart to avoid
it. The defrag code could be optimized to perform only 2 memcpy() calls.